rearrange docker image space source
